# WIP: Modify aggregate for efficiency