@nalimilan 2017-05-19
Looks nice. Can you explain (here and in the commit message) what the change consists in? Why would bypassing `combine` improve performance so much? Should `combine` be changed too?