@cjprybol 2017-05-24
I'm not sure how much of the efficiency comes from bypassing [`map`](https://github.com/JuliaData/DataTables.jl/blob/776f293cf85e92e6283a562bcc6cadb8bf006530/src/groupeddatatable/grouping.jl#L157) and how much comes from bypassing [`combine`](https://github.com/JuliaData/DataTables.jl/blob/776f293cf85e92e6283a562bcc6cadb8bf006530/src/groupeddatatable/grouping.jl#L194). `map` will construct `DataTables` for each `GroupApplied` which adds the overhead of the `DataTable` constructor multiple times. `combine` then `vcats` and `hcats` all of those `DataTables` together, adding additional checks, assertions, and overhead.

I don't want to touch `combine` here because I don't know how to support [anonymous functions that return `DataTables` as shown in `by` examples](https://github.com/JuliaData/DataTables.jl/blob/776f293cf85e92e6283a562bcc6cadb8bf006530/src/groupeddatatable/grouping.jl#L301) without it. If I can figure out how to support that with less overhead then we might be able to simplify some of this code more. I think it's safe to remove the calls to `map` and `combine` in aggregate because only one of the two `aggregate` methods used `map` and `combine`. With these changes, users will still have `by` for more complex split-apply-combine queries, and `aggregate` is a simpler and more efficient alternative.