@nalimilan 2017-10-08
@cjprybol I suppose this should be backported to DataFrames?