@cjprybol 2017-10-09
Yes, thanks for the reminder! I should also put these benchmarks into the beginnings of a PkgBenchmark suite. I think there are benchmarks we have scattered around other PRs in DataTables too. I couldn't figure out how to use PkgBenchmark the last time I tried, but [this comment](https://discourse.julialang.org/t/benchmarking-with-pkgbenchmark-jl/3038/13) covers most of the process