@cjprybol 2017-10-09
See https://github.com/JuliaData/DataFrames.jl/pull/1246