@nalimilan 2017-10-09
Cool. Note that PkgBenchmark has been largely refactored on git master, so probably better use that (see [this post](https://discourse.julialang.org/t/a-living-post-of-julia-vs-rs-data-manipulation-tasks-speeds/6316/16?u=nalimilan)).